[refactor] delete collection modal
